Moved Arch classes to separate module
